% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiate.R
\name{differentiate}
\alias{differentiate}
\title{Differenatiate a polynom}
\usage{
differentiate(P, differentation)
}
\arguments{
\item{P}{a \link[multipol]{multipol}}

\item{differentation}{integer vector, indicates the number of 
differentiations in each variable}
}
\value{
A multipol.
}
\description{
Differentiation of a multivariate polynomial.
}
\examples{
# define the polynomial 10*x^2*y*z^2
P <- as.multipol(array(0, dim=1+c(2,1,2)))
P[2,1,2] <- 10
# differentiate two times with respect to x
differentiate(P, c(2,0,0))
}
