% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evalPoly}
\alias{evalPoly}
\title{Evaluate a polynomial}
\usage{
evalPoly(P, Values)
}
\arguments{
\item{P}{a \link[multipol]{multipol}}

\item{Values}{a matrix of values at which the polynomial will be evaluate; 
\code{ncol(Values)} is the number of variables, \code{nrow(Values)} is the 
number of evaluation points}
}
\value{
A vector of length \code{nrow(Values)}.
}
\description{
Evaluation of a multivariate polynomial at some values of the 
variables.
}
\examples{
# define the polynomial 10*x^2*y*z^2
P <- as.multipol(array(0, dim=1+c(2,1,2)))
P[2,1,2] <- 10
# evaluate at (1,1,1) and (2,2,2):
evalPoly(P, rbind(c(1,1,1),c(2,2,2)))
}
