% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monomials.R
\docType{data}
\name{monomials}
\alias{monomials}
\alias{monomials}
\alias{xyz}
\alias{monomials}
\alias{x.}
\alias{monomials}
\alias{y.}
\alias{monomials}
\alias{z.}
\title{Monomials of three variables}
\format{An object of class \code{multipol} of dimension 2 x 1 x 1.}
\usage{
xyz(px, py, pz, coef)

x.

y.

z.
}
\arguments{
\item{coef}{coefficient of the monomial}

\item{pc, py, pz}{the powers of the monomial, integers}
}
\value{
A multipol.
}
\description{
Create monomials of three variables
}
\examples{
# create the monomial \\code{4*x*y^2*z^3}
xyz(1, 2, 3, 4)
# or
4*x.*y.^2*z.^3
}
\keyword{datasets}
