% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation.R
\name{representation}
\alias{representation}
\alias{representation}
\alias{representation}
\alias{representation2}
\alias{representation}
\alias{backrepresentation}
\alias{representation}
\alias{backrepresentation2}
\title{Representation of a polynomial}
\usage{
representation(P, degrees = NULL)

representation2(P)

backrepresentation(repr, monomials)

backrepresentation2(repr)
}
\arguments{
\item{P}{a multipol}

\item{degrees}{integer vector, the codings of the degrees of the monomials 
of \code{P} with non-zero coefficients; this is for internal purpose, set to 
\code{NULL} for an automatic derivation of this vector}

\item{repr}{a representation of a polynomial: a list with \code{degrees} 
and \code{coefficients}}

\item{monomials}{a matrix giving the degrees of the monomials forming the 
polynomial corresponding to \code{repr$degrees}}
}
\value{
The \code{representation} function returns a list with two elements: 
the coefficients in a numeric vector (\code{coefficients}) and an integer 
vector representing the monomial terms (\code{degrees}).
The \code{backrepresentation} function returns a \link[multipol]{multipol}.
}
\description{
Representation of a multivariate polynomial.
}
\note{
These functions, especially \code{backrepresentation}, are rather for 
internal purposes.
}
\examples{
P <- as.multipol(array(1:27, dim=c(3,3,3)))
representation(P)
}
